fix for file location issue, added BSD license
