+++ config/cam1.yaml
@@ -1,4 +1,4 @@
-serial_no: 32300451
+serial_no: 32301951
 cam_name: "camera1"
 yaml_url: "package://mcptam/calibrations/camera1.yaml"
 exposure_time: 30000
