+++ config/cam3.yaml
