+++ config/cam4.yaml
