+++ src/ximea_ros_cluster.cpp
@@ -1,4 +1,4 @@
-#include "ximea_ros_cluster.h"
+#include <ximea_camera/ximea_ros_cluster.h>
 
 int serial_nos[3] = { 32300651 ,  33300151 ,32301251};
 std::string cam_names[3] = {std::string("camera1"), std::string("camera2"), std::string("camera3")};
