+++ src/ximea_ros_driver.cpp
@@ -1,4 +1,4 @@
-#include "ximea_ros_driver.h"
+#include <ximea_camera/ximea_ros_driver.h>
 
 ximea_ros_driver::ximea_ros_driver(const ros::NodeHandle &nh, std::string cam_name, int serial_no, std::string yaml_url):ximea_driver(serial_no, cam_name){ 
 	pnh_ = nh;
