--- src/ximea_ros_driver.h
@@ -1,36 +1,0 @@
-#include "ximea_driver.h"
-#include <ros/ros.h>
-#include <image_transport/image_transport.h>
-#include <image_transport/publisher.h>
-#include <sensor_msgs/Image.h>
-#include <sensor_msgs/CameraInfo.h>
-#include <camera_info_manager/camera_info_manager.h>
-
-
-class ximea_ros_driver : public ximea_driver{
-	public:
-	ximea_ros_driver( const ros::NodeHandle &nh, std::string cam_name, int serial_no , std::string yaml_url);
-	ximea_ros_driver( const ros::NodeHandle &nh, std::string file_name);
-	//~ximea_ros_driver();
-	virtual void setImageDataFormat(std::string s);
-	void publishImage(const ros::Time & now);	//since these 2 functions should have the same time stamp we leave it up to the user to specify the timeif it is needed to do one or the other
-	void publishCamInfo(const ros::Time &now);
-	void publishImageAndCamInfo();
-	
-	protected:
-	ros::NodeHandle pnh_;
-	camera_info_manager::CameraInfoManager *cam_info_manager_;
-	image_transport::ImageTransport *it_;
-	image_transport::Publisher ros_cam_pub_;
-	ros::Publisher cam_info_pub_;
-
-	sensor_msgs::Image ros_image_;
-	sensor_msgs::CameraInfo cam_info_;
-	char * cam_buffer_;
-	int cam_buffer_size_;
-	int bpp_;		//the next 2 paramaeters are used by the ros_image_transport publisher
-	std::string encoding_;
-		
-	private:
-	void common_initialize(const ros::NodeHandle &nh);
-};
